Add more benchmarks

+++ src/cuda/mlperf/install_mlperf_and_virtual_environments.sh
@@ -14,20 +14,20 @@ else
     git clone https://github.com/mlcommons/inference.git
 fi
 
-cd $PWD/mlperf_inference/inference
+cd mlperf_inference/inference
 BASE_MLPERF_INFERENCE_DIR=$PWD
 
-if [ -f "$PWD/virtual_environment/bin/activate" ]; then
+if [ -f "$BASE_MLPERF_INFERENCE_DIR/virtual_environment_pytorch/bin/activate" ]; then
     echo "Virtual environment exists; skipping this part"
 else
-    wget https://gist.githubusercontent.com/vsajip/4673395/raw/0504ce930e6dc6b02e4955a07d91ad462e0ba80b/pyvenvex.py
+    wget -nc https://gist.githubusercontent.com/vsajip/4673395/raw/0504ce930e6dc6b02e4955a07d91ad462e0ba80b/pyvenvex.py
     #################################################
     # Single Pytorch environment for all workloads  #
     #################################################
-    python3 pyvenvex.py virtual_environment
+    python3 pyvenvex.py virtual_environment_pytorch
 fi
 
-. virtual_environment/bin/activate
+. virtual_environment_pytorch/bin/activate
 
 cd language/bert
 #python --version
@@ -54,7 +54,7 @@ else
 
     # Go back to bert, install dependencies and make
     cd ../language/bert
-    pip install onnx==1.8.1 transformers==2.4.0 onnxruntime==1.2.0 numpy==1.18.0
+    pip install onnx==1.8.1 transformers tokenization onnxruntime==1.2.0 numpy==1.18.0
     pip install torch===1.7.1+cu110 torchvision===0.8.2+cu110 torchaudio===0.7.2 -f https://download.pytorch.org/whl/torch_stable.html
     pip install nvidia-pyindex
     pip install nvidia-tensorflow
@@ -71,10 +71,10 @@ RESNET_DEPEND=$BASE_MLPERF_INFERENCE_DIR/resnet_inference_depend.done
 if [ -f "$RESNET_DEPEND" ]; then
     echo "RESNET dependencies fulfilled, skipping"
 else
-    wget https://zenodo.org/record/4588417/files/resnet50-19c8e357.pth .
-    wget https://zenodo.org/record/2535873/files/resnet50_v1.pb .
-    wget https://zenodo.org/record/4589637/files/resnet50_INT8bit_quantized.pt .
-    wget https://zenodo.org/record/2592612/files/resnet50_v1.onnx .
+    wget https://zenodo.org/record/4588417/files/resnet50-19c8e357.pth
+    wget https://zenodo.org/record/2535873/files/resnet50_v1.pb
+    wget https://zenodo.org/record/4589637/files/resnet50_INT8bit_quantized.pt
+    wget https://zenodo.org/record/2592612/files/resnet50_v1.onnx
     pip install opencv-python
     pip install pycocotools
     pip install future
@@ -117,65 +117,67 @@ else
     mkdir cocos-300-300
     mkdir cocos-1200-1200
     cd cocos-full
-    wget http://images.cocodataset.org/zips/val2017.zip
-    wget http://images.cocodataset.org/annotations/annotations_trainval2017.zip
+    wget -nc http://images.cocodataset.org/zips/val2017.zip
+    wget -nc http://images.cocodataset.org/annotations/annotations_trainval2017.zip
     unzip annotations_trainval2017.zip
     unzip val2017.zip
     cd $BASE_MLPERF_INFERENCE_DIR/tools/upscale_coco
     python upscale_coco.py --inputs $BASE_MLPERF_INFERENCE_DIR/vision/classification_and_detection/cocos/cocos-full --outputs $BASE_MLPERF_INFERENCE_DIR/vision/classification_and_detection/cocos/cocos-300-300 --size 300 300 --format jpg
     python upscale_coco.py --inputs $BASE_MLPERF_INFERENCE_DIR/vision/classification_and_detection/cocos/cocos-full --outputs $BASE_MLPERF_INFERENCE_DIR/vision/classification_and_detection/cocos/cocos-1200-1200 --size 1200 1200 --format jpg
     cd $BASE_MLPERF_INFERENCE_DIR/vision/classification_and_detection
-
+    python setup.py install
+    echo "SSD dependencies fulfilled" > $SSD_DEPEND
 fi
 
 deactivate
 
-: '
-##########################################
-# Setup the virtual environment for GNMT #
-##########################################
+################################################
+# Setup the Tensorflow 1.5 virtual environment #
+################################################
+
+cd $BASE_MLPERF_INFERENCE_DIR
+
 set -e
 
-cd ../../translation/gnmt/tensorflow
+wget -nc https://gist.githubusercontent.com/vsajip/4673395/raw/0504ce930e6dc6b02e4955a07d91ad462e0ba80b/pyvenvex.py
 
-wget https://gist.githubusercontent.com/vsajip/4673395/raw/0504ce930e6dc6b02e4955a07d91ad462e0ba80b/pyvenvex.py
+if [ -f "$BASE_MLPERF_INFERENCE_DIR/virtual_environment_tensorflow_1_15/bin/activate" ]; then
+    echo "Tensorflow 1.15 virtual env exists, skipping this part"
+else
+    python3 pyvenvex.py $BASE_MLPERF_INFERENCE_DIR/virtual_environment_tensorflow_1_15
+    . $BASE_MLPERF_INFERENCE_DIR/virtual_environment_tensorflow_1_15/bin/activate
+    pip install absl-py numpy 	
+    cd $BASE_MLPERF_INFERENCE_DIR/loadgen
+    pip install --force-reinstall dist/mlperf_loadgen-*.whl
+fi
 
-python3 pyvenvex.py virtual_environment
-. virtual_environment/bin/activate
+. $BASE_MLPERF_INFERENCE_DIR/virtual_environment_tensorflow_1_15/bin/activate
 
 set +e
 
-pip install absl-py numpy 	
-cd ../../loadgen
-pip install --force-reinstall dist/mlperf_loadgen-*.whl
+# GNMT
 
-cd ../translation/gnmt/tensorflow
+GNMT_DEPEND=$BASE_MLPERF_INFERENCE_DIR/gnmt_inference_depend.done
 
-# Install tensorflow 1.15
-# If we dont get the nvidia tensorflow
-# Ampere doesnt work
-
-pip install nvidia-pyindex
-pip install nvidia-tensorflow
-
-. download_trained_model.sh
-. download_dataset.sh
-#. verify_dataset.sh - Pausing this for now, checksums are screwy, blame zenodo
-
-# To run the benchmark
-# python run_task.py --run=performance --batch_size=32
+if [ -f "$GNMT_DEPEND" ]; then
+    echo "GNMT dependencies installed; skipping this part"
+else
+    cd $BASE_MLPERF_INFERENCE_DIR/translation/gnmt/tensorflow
+    pip install nvidia-pyindex
+    pip install nvidia-tensorflow
+    . download_trained_model.sh
+    . download_dataset.sh
+    echo "GNMT dependencies fulfilled" > $GNMT_DEPEND
+fi
 
 deactivate
 
-cd ../../../
-
-
-
+cd $BASE_MLPERF_INFERENCE_DIR/../../
 
 # Install the training workloads
 
 # cd ../../../../../
 
 . install_mlperf_and_virtual_environments_training.sh
-'
-echo "Done"
+
+echo 'Done'
