+++ src/cuda/mlperf/install_mlperf_and_virtual_environments_training.sh
@@ -3,65 +3,83 @@
 #   Install the training part    #
 ##################################
 
+echo "######################################"
+echo "#     Installing MLPerf training     #"
+echo "######################################"
+
 set -e
 
 mkdir mlperf_training
 cd mlperf_training
 git clone https://github.com/mlcommons/training.git
 
+BASE_MLPERF_DIR=$PWD/training
+
+cd training
+
 ##############################
 # RNN Translator
 ##############################
 
-cd rnn_translator
-wget https://gist.githubusercontent.com/vsajip/4673395/raw/0504ce930e6dc6b02e4955a07d91ad462e0ba80b/pyvenvex.py
+RNN_DEPEND=$BASE_MLPERF_INFERENCE_DIR/rnn_training_depend.done
 
-python3 pyvenvex.py virtual_environment
-. virtual_environment/bin/activate
+if [ -e "$BASE_MLPERF_DIR/virtual_environment_pytorch/bin/activate" ]; then
+    echo "Pytorch virtual env. exists, skipping"
+else
+    cd $BASE_MLPERF_DIR
+    wget https://gist.githubusercontent.com/vsajip/4673395/raw/0504ce930e6dc6b02e4955a07d91ad462e0ba80b/pyvenvex.py
 
-set +e
+    python3 pyvenvex.py virtual_environment
+    . virtual_environment/bin/activate
+fi
 
-pip install torch===1.7.1+cu110 -f https://download.pytorch.org/whl/torch_stable.html
-pip install sacrebleu
+RNN_DEPEND=$BASE_MLPERF_INFERENCE_DIR/rnn_training_depend.done
 
-. download_dataset.sh
+if [ -f "$RNN_DEPEND" ]; then
+    echo "RNN Translation dependencies fulfilled, skipping"
+else
+    set +e
+    pip install torch===1.7.1+cu110 -f https://download.pytorch.org/whl/torch_stable.html
+    pip install sacrebleu
 
-# To run this benchmark
-# python 
+    . download_dataset.sh
+    echo "RNN Translation dependencies fulfilled" > $RNN_DEPEND
+fi
 
 deactivate
 
 cd ..
-
-set -e
 
 ##############################
 # BERT
 ##############################
 
 # 365 GB of data
-# See if this is a problem
-cd language_model/tensorflow/bert
-wget https://gist.githubusercontent.com/vsajip/4673395/raw/0504ce930e6dc6b02e4955a07d91ad462e0ba80b/pyvenvex.py
+if [[ -z "${DOWNLOAD_BIG_BERT_TRAINING_DATASET}"]]; then
+    cd $BASE_MLPERF_DIR
+    set -e
 
-python3 pyvenvex.py virtual_environment
-. virtual_environment/bin/activate
+    wget https://gist.githubusercontent.com/vsajip/4673395/raw/0504ce930e6dc6b02e4955a07d91ad462e0ba80b/pyvenvex.py
 
-set +e
-pip install nvidia-pyindex
-pip install nvidia-tensorflow
+    python3 pyvenvex.py virtual_environment_tensorflow_1_15
+    . virtual_environment_tensorflow_1_15/bin/activate
 
-cd cleanup_scripts
-mkdir wiki
-cd wiki
-wget https://dumps.wikimedia.org/enwiki/20200101/enwiki-20200101-pages-articles-multistream.xml.bz2
-bzip2 -d enwiki-20200101-pages-articles-multistream.xml.bz2
-cd ..
-git clone https://github.com/attardi/wikiextractor.git
-python wikiextractor/WikiExtractor.py wiki/enwiki-20200101-pages-articles-multistream.xml
-. process_wiki '<text/*/wiki_??'
-python extract_test_set_articles.py
+    set +e
 
+    pip install nvidia-pyindex
+    pip install nvidia-tensorflow
+
+    cd cleanup_scripts
+    mkdir wiki
+    cd wiki
+    wget https://dumps.wikimedia.org/enwiki/20200101/enwiki-20200101-pages-articles-multistream.xml.bz2
+    bzip2 -d enwiki-20200101-pages-articles-multistream.xml.bz2
+    cd ..
+    git clone https://github.com/attardi/wikiextractor.git
+    python wikiextractor/WikiExtractor.py wiki/enwiki-20200101-pages-articles-multistream.xml
+    . process_wiki '<text/*/wiki_??'
+    python extract_test_set_articles.py
+fi
 
 deactivate
